+++ src/main/java/com/hubspot/jinjava/tree/ExpressionNode.java
@@ -21,7 +21,6 @@ import java.util.concurrent.TimeUnit;
 import org.apache.commons.lang3.StringUtils;
 
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
-import com.hubspot.jinjava.interpret.UnknownTokenException;
 import com.hubspot.jinjava.lib.filter.EscapeFilter;
 import com.hubspot.jinjava.tree.output.OutputNode;
 import com.hubspot.jinjava.tree.output.RenderedOutputNode;
@@ -45,10 +44,6 @@ public class ExpressionNode extends Node {
     StringBuilder sb = new StringBuilder();
 
     Object var = interpreter.resolveELExpression(master.getExpr(), getLineNumber());
-
-    if (var == null && interpreter.getConfig().isFailOnUnknownTokens()) {
-      throw new UnknownTokenException(master.getExpr(), getLineNumber());
-    }
 
     String result = Objects.toString(var, "");