+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Jinjava Releases #
 
+### 2017-08-12 Version 2.2.8 ([Maven Central](https://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.2.8%22)) ###
+
+* Prevent recursion in Jinjava.
+* Fix failsOnUnknownTokens.
+* Add EscapeJson filter.
+
 ### 2017-08-12 Version 2.2.7 ([Maven Central](https://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.2.7%22)) ###
 
 * Delegate toString() method on PyMap